[Refactor] jsonのパース処理を行う関数を独立したファイルへ分離
nlohmann-json が不要にインクルードされるのを避けるため、jsonのパースを
行う関数を独立したファイルへ分離する。
