+++ src/info-reader/info-reader-util.cpp
@@ -1,5 +1,6 @@
 #include "info-reader/info-reader-util.h"
 #include "artifact/random-art-effects.h"
+#include "info-reader/parse-error-types.h"
 #include "main/angband-headers.h"
 #include "object-enchant/activation-info-table.h"
 #include "util/enum-converter.h"
@@ -59,50 +60,6 @@ void append_english_text(std::string &text, std::string_view add)
 #endif
 
 /*!
- * @brief JSON Objectから文字列をセットする
- *
- * 引数で与えられたJSON Objectから日本語版の場合は"ja"、英語版の場合は"en"のキーで文字列を取得し、
- * data に格納する。キーが存在しない場合は is_required が真の場合はエラーを返し、偽の場合は何もせずに終了する。
- *
- * @param json 文字列の格納されたJSON Object
- * @param data 文字列を格納する変数への参照
- * @param is_required 必須かどうか
- * @return エラーコード
- */
-errr info_set_string(const nlohmann::json &json, std::string &data, bool is_required)
-{
-    const auto unexist_result = is_required ? PARSE_ERROR_TOO_FEW_ARGUMENTS : PARSE_ERROR_NONE;
-
-    if (json.is_null()) {
-        return unexist_result;
-    }
-
-    if (!json.is_object()) {
-        return PARSE_ERROR_TOO_FEW_ARGUMENTS;
-    }
-
-#ifdef JP
-    const auto &ja_str = json.find("ja");
-    if (ja_str == json.end()) {
-        return unexist_result;
-    }
-    auto ja_str_sys = utf8_to_sys(ja_str->get<std::string>());
-    if (!ja_str_sys) {
-        return PARSE_ERROR_INVALID_FLAG;
-    }
-    data = std::move(*ja_str_sys);
-#else
-    const auto &en_str = json.find("en");
-    if (en_str == json.end()) {
-        return unexist_result;
-    }
-    data = en_str->get<std::string>();
-#endif
-
-    return PARSE_ERROR_NONE;
-}
-
-/*!
  * @brief ダイスを表す文字列を解析してダイスの値をセットする
  *
  * 引数で与えられた文字列 "XdY" をダイスの値に変換し、X を dd に、Y を ds に格納する。
@@ -113,7 +70,7 @@ errr info_set_string(const nlohmann::json &json, std::string &data, bool is_requ
  * @param ds ダイスの面数を格納する変数への参照
  * @return エラーコード
  */
-errr info_set_dice(const std::string_view dice_str, DICE_NUMBER &dd, DICE_SID &ds)
+errr info_set_dice(std::string_view dice_str, DICE_NUMBER &dd, DICE_SID &ds)
 {
     const auto &dice = str_split(dice_str, 'd', false, 2);
     if (dice.size() < 2) {
@@ -124,22 +81,4 @@ errr info_set_dice(const std::string_view dice_str, DICE_NUMBER &dd, DICE_SID &d
     ds = std::stoi(dice[1]);
 
     return PARSE_ERROR_NONE;
-}
-
-/*!
- * @brief JSON Objectからダイスの値を取得する
- * @param json ダイスの値が格納されたJSON Object
- * @param dd ダイスの数を格納する変数への参照
- * @param ds ダイスの面数を格納する変数への参照
- * @param is_required 必須かどうか
- * 必須でJSON Objectがnullや文字列でない場合はエラーを返す。必須でない場合は何もせずに終了する。
- * @return エラーコード
- */
-errr info_set_dice(const nlohmann::json &json, DICE_NUMBER &dd, DICE_SID &ds, bool is_required)
-{
-    if (json.is_null() || !json.is_string()) {
-        return is_required ? PARSE_ERROR_TOO_FEW_ARGUMENTS : PARSE_ERROR_NONE;
-    }
-
-    return info_set_dice(json.get<std::string>(), dd, ds);
 }
