+++ src/info-reader/info-reader-util.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#include "external-lib/include-json.h"
-#include "info-reader/parse-error-types.h"
-#include "locale/japanese.h"
 #include "system/angband.h"
 #include "util/bit-flags-calculator.h"
 #include <concepts>
@@ -25,9 +22,7 @@ RandomArtActType grab_one_activation_flag(std::string_view what);
 void append_english_text(std::string &text, std::string_view add);
 #endif
 
-errr info_set_string(const nlohmann::json &json, std::string &data, bool is_required);
-errr info_set_dice(const std::string_view dice_str, DICE_NUMBER &dd, DICE_SID &ds);
-errr info_set_dice(const nlohmann::json &json, DICE_NUMBER &dd, DICE_SID &ds, bool is_required);
+errr info_set_dice(std::string_view dice_str, DICE_NUMBER &dd, DICE_SID &ds);
 
 /// @note clang-formatによるconceptの整形が安定していないので抑制しておく
 // clang-format off
@@ -104,40 +99,4 @@ template <typename T>
 void info_set_value(T &arg, const std::string &val, int base = 10)
 {
     arg = static_cast<T>(std::stoi(val, nullptr, base));
-}
-
-using Range = std::pair<int, int>;
-
-template <typename T>
-concept IntegralOrEnum = std::integral<T> || std::is_enum_v<T>;
-
-/*!
- * @brief JSON Objectから整数値もしくはenum値を取得する
-
- * 引数で与えられたJSON Objectから整数値もしくはenum値を取得し、 data に格納する。
- *
- * @param json 整数値が格納されたJSON Object
- * @param data 値を格納する変数への参照
- * @param is_required 必須かどうか。
- * 必須でJSON Objectがnullの場合はエラーを返す。必須でない場合は何もせずに終了する。
- * @param range 取得した値の範囲を指定する。範囲外の値が取得された場合はエラーを返す。
- * @return エラーコード
- */
-template <IntegralOrEnum T>
-errr info_set_integer(const nlohmann::json &json, T &data, bool is_required, std::optional<Range> range = std::nullopt)
-{
-    if (json.is_null()) {
-        return is_required ? PARSE_ERROR_TOO_FEW_ARGUMENTS : PARSE_ERROR_NONE;
-    }
-    if (!json.is_number_integer()) {
-        return PARSE_ERROR_TOO_FEW_ARGUMENTS;
-    }
-
-    const auto value = json.get<T>();
-    if (range && (value < static_cast<T>(range->first) || value > static_cast<T>(range->second))) {
-        return PARSE_ERROR_INVALID_FLAG;
-    }
-
-    data = value;
-    return PARSE_ERROR_NONE;
 }
