+++ src/info-reader/race-reader.cpp
@@ -2,6 +2,7 @@
 #include "alliance/alliance.h"
 #include "artifact/fixed-art-types.h"
 #include "info-reader/info-reader-util.h"
+#include "info-reader/json-reader-util.h"
 #include "info-reader/parse-error-types.h"
 #include "info-reader/race-info-tokens-table.h"
 #include "locale/japanese.h"
